+++ src/Client.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Gianfriaur\OpcuaPhpClient;
 
 use Gianfriaur\OpcuaPhpClient\Client\ManagesAutoRetryTrait;
+use Gianfriaur\OpcuaPhpClient\Client\ManagesBatchingTrait;
 use Gianfriaur\OpcuaPhpClient\Client\ManagesBrowseTrait;
 use Gianfriaur\OpcuaPhpClient\Client\ManagesConnectionTrait;
 use Gianfriaur\OpcuaPhpClient\Client\ManagesHandshakeTrait;
@@ -38,6 +39,7 @@ class Client implements OpcUaClientInterface
 {
     use ManagesTimeoutTrait;
     use ManagesAutoRetryTrait;
+    use ManagesBatchingTrait;
     use ManagesConnectionTrait;
     use ManagesHandshakeTrait;
     use ManagesSecureChannelTrait;